+++ src/App.jsx
@@ -1,10 +1,10 @@
 import './App.css';
-import { SortingTable } from './components/SortingTable';
+import { FilteringTable } from './components/FilteringTable';
 
 function App() {
   return (
     <div className='App'>
-      <SortingTable />
+      <FilteringTable />
     </div>
   );
 }
