+++ src/components/FilteringTable.jsx
@@ -1,0 +1,70 @@
+import { useTable, useGlobalFilter } from 'react-table';
+import MOCK_DATA from './MOCK_DATA.json';
+import { COLUMNS } from './columns';
+import { useMemo } from 'react';
+import './table.css';
+import { GlobalFilter } from './GlobalFilter';
+
+export const FilteringTable = () => {
+  const columns = useMemo(() => COLUMNS, []);
+  const data = useMemo(() => MOCK_DATA, []);
+
+  const {
+    getTableProps,
+    getTableBodyProps,
+    headerGroups,
+    footerGroups,
+    rows,
+    prepareRow,
+    state,
+    setGlobalFilter,
+  } = useTable(
+    {
+      columns,
+      data,
+    },
+    useGlobalFilter
+  );
+
+  const { globalFilter } = state;
+
+  return (
+    <>
+      <GlobalFilter filter={globalFilter} setFilter={setGlobalFilter} />
+      <table {...getTableProps()}>
+        <thead>
+          {headerGroups.map((headerGroup) => (
+            <tr {...headerGroup.getHeaderGroupProps()}>
+              {headerGroup.headers.map((column) => (
+                <th {...column.getHeaderProps()}>{column.render('Header')}</th>
+              ))}
+            </tr>
+          ))}
+        </thead>
+        <tbody {...getTableBodyProps()}>
+          {rows.map((row) => {
+            prepareRow(row);
+            return (
+              <tr {...row.getRowProps()}>
+                {row.cells.map((cell) => {
+                  return (
+                    <td {...cell.getCellProps()}>{cell.render('Cell')}</td>
+                  );
+                })}
+              </tr>
+            );
+          })}
+        </tbody>
+        <tfoot>
+          {footerGroups.map((footerGroups) => (
+            <tr {...footerGroups.getFooterGroupProps()}>
+              {footerGroups.headers.map((column) => (
+                <td {...column.getFooterProps}>{column.render('Footer')}</td>
+              ))}
+            </tr>
+          ))}
+        </tfoot>
+      </table>
+    </>
+  );
+};
