--- src/components/SortingTable.jsx
@@ -1,65 +1,0 @@
-import { useTable, useSortBy } from 'react-table';
-import MOCK_DATA from './MOCK_DATA.json';
-import { COLUMNS } from './columns';
-import { useMemo } from 'react';
-import './table.css';
-
-export const SortingTable = () => {
-  const columns = useMemo(() => COLUMNS, []);
-  const data = useMemo(() => MOCK_DATA, []);
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    footerGroups,
-    rows,
-    prepareRow,
-  } = useTable(
-    {
-      columns,
-      data,
-    },
-    useSortBy
-  );
-
-  return (
-    <table {...getTableProps()}>
-      <thead>
-        {headerGroups.map((headerGroup) => (
-          <tr {...headerGroup.getHeaderGroupProps()}>
-            {headerGroup.headers.map((column) => (
-              <th {...column.getHeaderProps(column.getSortByToggleProps())}>
-                {column.render('Header')}
-                <span>
-                  {column.isSorted ? (column.isSortedDesc ? ' 🔽' : ' 🔼') : ''}
-                </span>
-              </th>
-            ))}
-          </tr>
-        ))}
-      </thead>
-      <tbody {...getTableBodyProps()}>
-        {rows.map((row) => {
-          prepareRow(row);
-          return (
-            <tr {...row.getRowProps()}>
-              {row.cells.map((cell) => {
-                return <td {...cell.getCellProps()}>{cell.render('Cell')}</td>;
-              })}
-            </tr>
-          );
-        })}
-      </tbody>
-      <tfoot>
-        {footerGroups.map((footerGroups) => (
-          <tr {...footerGroups.getFooterGroupProps()}>
-            {footerGroups.headers.map((column) => (
-              <td {...column.getFooterProps}>{column.render('Footer')}</td>
-            ))}
-          </tr>
-        ))}
-      </tfoot>
-    </table>
-  );
-};
